api calls are maded in photos
