--- src/components/Favourites/Favourites.css
@@ -1,21 +1,0 @@
-.favourites {
-  margin-bottom: 20px;
-  display: grid;
-}
-
-.favourite-images {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  border-radius: 0;
-}
-
-.fav-image {
-  width: 100%;
-}
-
-.fav-image img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
